add missed custom pokt headers at mainnet http provider
